requirements.txt
celery==5.6.2
lxml==6.0.2
playwright==1.58.0
python-dotenv==1.2.1
SQLAlchemy==2.0.46
